databricks/requirements.txt
azure-ai-formrecognizer>=3.3.0
azure-storage-blob>=12.19.0
azure-search-documents>=11.4.0
azure-ai-inference>=1.0.0b6
tiktoken>=0.7.0
